rs: remove delegated authorization provider
This became very problematic when trying to use the client across
multiple threads; it requires introducing tons of additional constraints
on the futures and I couldn't get it into a place that was workable.

I think it's reasonable for the caller to handle dealing with
authorization. For instance, by having a "bearer" class is called to
return the client and automatically re-acquires credentials at that
time if necessary.