Do not check host RemoteForwarders when PFS channel opens
There is a small time interval in ForwardPortAsync() between the port being forwarded with ForwardFromRemotePortAsync() and RemoteForwarders updated.

See https://github.com/microsoft/basis-planning/issues/467

The check is redundant because setting PFS.AcceptRemoteConnectionsForNonForwardedPorts to false makes PFS reject forwarding requests from the clients for the ports that are not forwarded and are missing in PFS.remoteConnectors.
Call to PFS.ForwardFromRemotePortAsync() in ForwardPortAsync() adds the connector to PFS.remoteConnectors.