rs: implement initial host relay  (#127)
* rs: allow rebuilding the client with new options

We use this in VS Code to derive a client that uses host tokens from
the "base" client that users user auth.

* rs: implement initial host relay

This implement a host relay. It does not yet implement reconnections,
although with that added in consumer code this has been working fairly
well for me. The relay is opt-in under the `connections` feature flag.

* use published ruush versions

* improve logging, key exchange, pk handling

* address pr comments

* fix readbuffer dropping bytes

* one last io fix...

* update russh

* fix: temporarily use vscode build for russh

* address some pr comments