Reconnect tunnel client on 503
Fix for https://github.com/microsoft/basis-planning/issues/581

Normally nginx choses another healthy pod when it encounters 503.
However, if there are no other pods, it returns 503 to the client.
This rare case may happen when the cluster recovers from a failure
and the nginx controller has started but Relay service has not yet.