Fix TunnelConnectionSession.getFreshAccessToken (#219)
Fix for https://github.com/microsoft/basis-planning/issues/806

When refreshing tunnel access token, `TunnelConnectionSession.gGetFreshAccessToken()` calls `tunnelManager` if it was set to fetch the current tunnel using `tunnel` property. If tunnel has a valid access token, then fetching succeeds but the fetched tunnel may not have the access token, and so a valid access token on `this.tunnel` object may be ignored.

The fix is to use the access token from `this.tunnel` if the fetched tunnel doesn't have one.